+++ data/Pandora.Definitions.MicrosoftGraph.Beta/CommonTypes/Model-SecurityExportFileMetadata.cs
@@ -23,5 +23,5 @@ internal class SecurityExportFileMetadataModel
     public string? ODataType { get; set; }
 
     [JsonPropertyName("size")]
-    public long? Size { get; set; }
+    public int? Size { get; set; }
 }
